+++ Riot/Modules/Application/LegacyAppDelegate.m
@@ -1107,7 +1107,6 @@ NSString *const AppDelegateUniversalLinkDidChangeNotification = @"AppDelegateUni
 
 - (void)pushNotificationService:(PushNotificationService *)pushNotificationService shouldNavigateToRoomWithId:(NSString *)roomId
 {
-    _lastNavigatedRoomIdFromPush = roomId;
     [self navigateToRoomById:roomId];
 }